agriscan/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.12
pillow==11.1.0
torch
torchvision
